Changed names of findAllValidMoves and removeDuplicateMoves to findAllValidCombos and removeDuplicateCombos respectively, to better reflect their purpose.
Added requirement for adding the second move of a combo, that the combined distance travelled by the firstMove and the secondMove must be equal to the sum of dice1 and dice2, i.e. if you roll a 1 and a 3, it will no longer allow combos like "1-2, 5-6".

Implemented findBarMoves() and separately added code to findAllValidCombos to remove any remaining barMoves from copyAllMoves if firstMove would have already used up all the pieces on the bar.