Fixed original code to add "follow-on" moves to copyAllMoves.
Added code to remove combos from allCombos if they have less moves than the combo with the most moves, i.e. if you *can* make a play with 2 moves, then you have to, but if there are none, then combos of only one move would be included and could be played. Same for 4 and 3.