Changed where copyAllMove2 and copyAllMove3 are assigned so that only valid 4-move combos should show up now, not combos where later moves were included that wrongly relied on an assumption of a certain previous move happening earlier in the same combo.
Fixed separate issue where later moves starting on the same strip as firstMove were always removed, when it's supposed to be that they're only removed if that strip only has one piece on it to begin with (or 2 pieces to begin with for thirdMoves, or 3 for fourthMoves).

To recap, the list no longer has any invalid plays, but still does have duplicates, and as far as I can tell is no longer missing any plays where only the board is involved, but likely is when the bar and bearOff get involved.